last redeemToDenom in zeroEx
