+++ jest.config.ts
@@ -31,10 +31,11 @@ const projects = [
   // project('core', ['tests/SelfTest/test-streaming']),
   // project('core', ['tests/SelfTest/test-feemanager-stream']),
   // project('core', ['tests/SelfTest/test-vault']),
-  project('core', ['tests/SelfTest/test-zeroAdapter']),
+  // project('core', ['tests/SelfTest/test-zeroAdapter']),
+  project('core', ['tests/SelfTest/test-zeroAdapter-takeorder']),
 ].filter((project) => !!project);
 
 export default {
   projects,
-  testTimeout: 1200000,
+  testTimeout: 12000000,
 };
