+++ package.json
@@ -29,8 +29,8 @@
     "format:solhint": "yarn lint:solhint --fix",
     "fork:mainnet": "hardhat node --fork https://mainnet.infura.io/v3/b8af1adb9cbe494cb0be1611d283e4c7",
     "fork:kovan": "hardhat node --fork https://kovan.infura.io/v3/b8af1adb9cbe494cb0be1611d283e4c7",
-    "fork:alkovan": "hardhat node --fork https://eth-kovan.alchemyapi.io/v2/-BPEIJTGZJkx7WbtkL89BYsF-_t3Scwg",
-    "fork:almainnet": "hardhat node --fork https://eth-mainnet.alchemyapi.io/v2/-BPEIJTGZJkx7WbtkL89BYsF-_t3Scwg",
+    "fork:alkovan": "hardhat node --fork https://eth-kovan.alchemyapi.io/v2/-BPEIJTGZJkx7WbtkL89BYsF-_t3Scwg --fork-block-number 27277405",
+    "fork:almainnet": "hardhat node --fork https://eth-mainnet.alchemyapi.io/v2/-BPEIJTGZJkx7WbtkL89BYsF-_t3Scwg --fork-block-number 12540501",
     "test": "jest",
     "test:watch": "NODE_OPTIONS=TS_JEST_DISABLE_VER_CHECKER=true jest --watchAll",
     "test:coverage": "jest --coverage",
